drm/tidss: Rename "hw_videoport" to "vp_idx"
The DSS outputs DPI video signals through its Video Ports (VPs), and
based on the SoC version, a DSS can have as many as 4 VPs.
"hw_videoport" was used to refer the index of the VP in question, but
that can become confusing with the output ports available on the SoC for
displaying data.

Rename "hw_videoport" to "vp_idx" in order to remove the clarify.

Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>