drm/bridge: cdns-dsi: Add support to send zero parameter DCS commands
If any normal DCS write command has already been transmitted prior to
transmitting any Zero-Parameter DCS command, then it is necessary to
clear the TX FIFO by resetting it. Otherwise, the FIFO points to another
location, and the DCS command transmits unnecessary data causing the
panel to not work[0].

Further, to transmit the zero parameter command, zero must be written to
the DIRECT_CMD_WRDATA register.

Add support for this.

[0]: Section 12.6.5.7.5.2: "Command Mode Settings" in TDA4VM Technical
     Reference Manual: https://www.ti.com/lit/zip/spruil1

Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>