drm/bridge: panel: Fix the device to which panel-bridge is registered
The panel-bridge resource is getting registered to the device of the
preceding bridge that calls for the panel-bridge creation. But the
memory alloc of the panel-bridge is registered to the device of the
panel itself.

Fix it so that the panel-bridge resource also gets allocated to the
device of the panel, and not that of the preceding bridge.

Fixes: 67022227ffb1 ("drm/bridge: Add a devm_ allocator for panel bridge.")
Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>