drm/bridge: cdns-dsi: Fix phy de-init and flag it so
The driver code doesn't have a de-initialization path as yet, and so it
does not clear the phy_initialized flag while suspending. This is a
problem because after resume the driver looks at this flag to determine
if a Phy re-initialization is required or not. It is in fact required
because the hardware is resuming from a suspend, but the driver does not
carry out any re-initialization causing the D-Phy to not work at all.

Move the Phy initialization from _bridge_enable() to _resume(), and
de-initialize during the _suspend() - so that the phy_{init, exit} take
place once every resume/suspend cycle.

Also add the counterpart of phy_power_on(), that is phy_power_off() from
the _bridge_disable() and clear the flags so that the Phy can be
initialized again when required.

The order of calls still remains the same. phy_init() needs to be called
before phy_power_on() - which happens still. What this patch changes is
the frequency of the phy_init() call. Instead of it being called once
every bridge enable/disable cycle, it is now being called once every
resume/suspend cycle.

Fixes: fced5a364dee ("drm/bridge: cdns: Convert to phy framework")
Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>
Signed-off-by: Aradhya Bhatia <aradhya.bhatia@linux.dev>